# api-contract.yaml
openapi: 3.1.0
info:
  title: Routa.js API Contract
  description: |
    Single source of truth for the Routa.js dual-backend API.
    Both the Next.js backend (src/app/api/) and the Rust backend (crates/routa-server/)
    MUST implement all endpoints defined here with compatible request/response shapes.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Next.js backend (dev)
  - url: http://localhost:3210
    description: Rust backend (desktop)

# ─────────────────────────────────────────────────────────
# Shared component schemas
# ─────────────────────────────────────────────────────────
components:
  schemas:
...
    AgentRole:
      type: string
      enum: [ROUTA, CRAFTER, GATE, DEVELOPER]
    ModelTier:
      type: string
      enum: [SMART, BALANCED, FAST]
    AgentStatus:
      type: string
      enum: [PENDING, ACTIVE, COMPLETED, ERROR, CANCELLED]
    TaskStatus:
      type: string
      enum: [PENDING, IN_PROGRESS, REVIEW_REQUIRED, COMPLETED, NEEDS_FIX, BLOCKED, CANCELLED]
...
    VerificationVerdict:
      type: string
      enum: [APPROVED, NOT_APPROVED, BLOCKED]
...
    TaskPriority:
      type: string
      enum: [low, medium, high, urgent]
...
    Agent:
      type: object
      required: [id, name, role, modelTier, workspaceId, status, createdAt, updatedAt, metadata]
      properties:
        id:
          type: string
        name:
          type: string
        role:
          $ref: "#/components/schemas/AgentRole"
        modelTier:
          $ref: "#/components/schemas/ModelTier"
        workspaceId:
          type: string
        parentId:
          type: string
        status:
          $ref: "#/components/schemas/AgentStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties:
            type: string
    Task:
      type: object
      required: [id, title, objective, status, dependencies, workspaceId, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        objective:
          type: string
        scope:
          type: string
        acceptanceCriteria:
          type: array
          items:
            type: string
        verificationCommands:
          type: array
          items:
            type: string
        assignedTo:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        boardId:
          type: string
        columnId:
          type: string
        position:
          type: integer
        priority:
          $ref: "#/components/schemas/TaskPriority"
        labels:
          type: array
          items:
            type: string
        assignee:
          type: string
        assignedProvider:
          type: string
        assignedRole:
          $ref: "#/components/schemas/AgentRole"
        assignedSpecialistId:
          type: string
        assignedSpecialistName:
          type: string
        triggerSessionId:
          type: string
        vcsId:
          type: string
        vcsNumber:
          type: integer
        vcsUrl:
          type: string
        vcsRepo:
          type: string
        vcsState:
          type: string
        vcsSyncedAt:
          type: string
          format: date-time
        lastSyncError:
          type: string
        dependencies:
          type: array
          items:
            type: string
        parallelGroup:
          type: string
        workspaceId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completionSummary:
          type: string
        verificationVerdict:
          $ref: "#/components/schemas/VerificationVerdict"
        verificationReport:
          type: string
...
    DeletedResponse:
      type: object
      required: [deleted]
      properties:
        deleted:
          type: boolean
          const: true
...
paths:
...
  /api/agents:
    get:
      operationId: listAgents
      summary: List agents (or get single by id query param)
      parameters:
        - name: id
          in: query
          schema:
            type: string
          description: If provided, returns single agent
        - name: workspaceId
          in: query
          schema:
            type: string
            default: "default"
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/AgentRole"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/AgentStatus"
        - name: parentId
          in: query
          schema:
            type: string
        - name: summary
          in: query
          schema:
            type: string
          description: If present with id, returns agent summary (Next.js only)
      responses:
        "200":
          description: Agent list or single agent
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      agents:
                        type: array
                        items:
                          $ref: "#/components/schemas/Agent"
                  - $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      summary: Create a new agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, role]
              properties:
                name:
                  type: string
                role:
                  $ref: "#/components/schemas/AgentRole"
                workspaceId:
                  type: string
                  default: "default"
                parentId:
                  type: string
                modelTier:
                  $ref: "#/components/schemas/ModelTier"
      responses:
        "200":
          description: Created agent
          content:
            application/json:
              schema:
                type: object
                properties:
                  agentId:
                    type: string
                  agent:
                    $ref: "#/components/schemas/Agent"
...
  /api/tasks:
    get:
      operationId: listTasks
      summary: List tasks
      parameters:
        - name: workspaceId
          in: query
          schema:
            type: string
            default: "default"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - name: assignedTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, objective]
              properties:
                title:
                  type: string
                objective:
                  type: string
                workspaceId:
                  type: string
                  default: "default"
                scope:
                  type: string
                acceptanceCriteria:
                  type: array
                  items:
                    type: string
                verificationCommands:
                  type: array
                  items:
                    type: string
                dependencies:
                  type: array
                  items:
                    type: string
                parallelGroup:
                  type: string
                boardId:
                  type: string
                columnId:
                  type: string
                position:
                  type: integer
                priority:
                  $ref: "#/components/schemas/TaskPriority"
                labels:
                  type: array
                  items:
                    type: string
                assignee:
                  type: string
                assignedProvider:
                  type: string
                assignedRole:
                  $ref: "#/components/schemas/AgentRole"
                assignedSpecialistId:
                  type: string
                assignedSpecialistName:
                  type: string
                createGitHubIssue:
                  type: boolean
                repoPath:
                  type: string
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteAllTasks
      summary: Delete all tasks for a workspace
      parameters:
        - name: workspaceId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
  /api/tasks/{id}:
    get:
      operationId: getTask
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Single task
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                objective:
                  type: string
                status:
                  $ref: "#/components/schemas/TaskStatus"
                boardId:
                  type: string
                columnId:
                  type: string
                position:
                  type: integer
                priority:
                  $ref: "#/components/schemas/TaskPriority"
                labels:
                  type: array
                  items:
                    type: string
                assignee:
                  type: string
                assignedProvider:
                  type: string
                assignedRole:
                  $ref: "#/components/schemas/AgentRole"
                assignedSpecialistId:
                  type: string
                assignedSpecialistName:
                  type: string
                repoPath:
                  type: string
                syncToGitHub:
                  type: boolean
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
    delete:
      operationId: deleteTask
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"
...
  /api/clone/branches:
    get:
      operationId: getBranches
      summary: Get branch info
      parameters:
        - name: repoPath
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Branch info
    post:
      operationId: fetchBranches
      summary: Fetch remote branches
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                repoPath:
                  type: string
      responses:
        "200":
          description: Fetched
    patch:
      operationId: checkoutBranch
      summary: Checkout branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                repoPath:
                  type: string
                branch:
                  type: string
                pull:
                  type: boolean
      responses:
        "200":
          description: Checked out
    delete:
      operationId: deleteBranch
      summary: Delete local branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                repoPath:
                  type: string
                branch:
                  type: string
      responses:
        "200":
          description: Deleted
...
  /api/a2a/message:
    post:
      operationId: sendA2AMessage
      summary: Send a message via the A2A protocol
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Message sent
...